+++ packages/health/CHANGELOG.md
@@ -6,6 +6,8 @@
 * Updated `intl` to ^0.20.1 - Closes [#1092](https://github.com/cph-cachet/flutter-plugins/issues/1092)
 * Updated `device_info_plus` to ^11.2.0
 * Example app: Updated `permission_handler` to ^11.3.1
+* iOS: Add `WATER_TEMPERATURE` and `UNDERWATER_DEPTH` health values - Closes [#1096](https://github.com/cph-cachet/flutter-plugins/issues/1096)
+* iOS: Add support for `Underwater Diving` workout - Closes [#1096](https://github.com/cph-cachet/flutter-plugins/issues/1096)
 
 ## 11.1.1